add ScraperScheduler
